backend/requirements.txt
# Core Flask and extensions
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Security-Too==5.6.2
Flask-Principal==0.4.0
Flask-WTF==1.2.2
Flask-Admin==1.6.1
flask-cors==6.0.1

# API Documentation (Flask-RESTX instead of flasgger/flask-swagger)
flask-restx==1.3.0

# Database
alembic==1.16.4
SQLAlchemy==2.0.42

# Security and Authentication
bcrypt==4.3.0
PyJWT==2.10.1

# Utilities
python-dotenv==1.1.1
email_validator==2.2.0
requests==2.32.3

# Template engine and markup
Jinja2==3.1.6
MarkupSafe==3.0.2
WTForms==3.0.1

# WSGI Server for production
gunicorn==23.0.0

# Core Python utilities
click==8.2.1
blinker==1.9.0
itsdangerous==2.2.0
Werkzeug==3.1.3

# Validation and serialization
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
referencing==0.36.2
rpds-py==0.27.0

# Networking and HTTP
certifi==2025.8.3
charset-normalizer==3.4.3
dnspython==2.7.0
idna==3.10
urllib3==2.5.0

# Migration and template support
Mako==1.3.10
mistune==3.1.3
packaging==25.0
PyYAML==6.0.2
six==1.17.0

# Python compatibility
typing_extensions==4.14.1
importlib_resources==6.5.2
greenlet==3.2.4